modify entity user,review 31/3

+++ database/src/main/java/ut/edu/database/repositories/ReviewRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     List<Review> findByPropertyId(Long propertyId);
     List<Review> findByUserId(Long userId);
-    List<Review> findByRating(int rating);
+    List<Review> findByRating(Byte rating);
 }
